update handle create and delete a user

+++ src/main/java/vn/hoidanit/jobhunter/controller/UserController.java
@@ -4,10 +4,13 @@ import org.springframework.web.bind.annotation.RestController;
 
 import com.turkraft.springfilter.boot.Filter;
 
+import jakarta.validation.Valid;
 import vn.hoidanit.jobhunter.domain.User;
+import vn.hoidanit.jobhunter.domain.dto.ResCreateUserDTO;
 import vn.hoidanit.jobhunter.domain.dto.ResultPaginationDTO;
 import vn.hoidanit.jobhunter.service.UserService;
 import vn.hoidanit.jobhunter.util.anotation.APIMessage;
+import vn.hoidanit.jobhunter.util.error.IdInvaliException;
 
 import org.springframework.data.domain.Pageable;
 import org.springframework.data.jpa.domain.Specification;
@@ -35,11 +38,18 @@ public class UserController {
     }
 
     @PostMapping("/users")
-    public ResponseEntity<User> createNewUser(@RequestBody User userinput) {
+    @APIMessage("Create a new user")
+    public ResponseEntity<ResCreateUserDTO> createNewUser(@Valid @RequestBody User userinput) throws IdInvaliException {
+        boolean isEmailExist = this.userService.isEmailExist(userinput.getEmail());
+
+        if (isEmailExist) {
+            throw new IdInvaliException("Email " + userinput.getEmail() + " is exist, please input another email!");
+        }
+
         String hashPassword = this.passwordEncoder.encode(userinput.getPassword());
         userinput.setPassword(hashPassword);
         User newUser = this.userService.handleCreateUser(userinput);
-        return ResponseEntity.status(HttpStatus.CREATED).body(newUser);
+        return ResponseEntity.status(HttpStatus.CREATED).body(this.userService.ConvertToResCreateDTO(newUser));
     }
 
     @GetMapping("/users")
@@ -61,7 +71,13 @@ public class UserController {
     }
 
     @DeleteMapping("/users/{id}")
-    public ResponseEntity<Void> deletUser(@PathVariable("id") long id) {
+    @APIMessage("Delete a user")
+    public ResponseEntity<Void> deletUser(@PathVariable("id") long id) throws IdInvaliException {
+        User currentUser = this.userService.fetchUserById(id);
+        if (currentUser == null) {
+            throw new IdInvaliException("User with ID = " + id + " do not exist!");
+        }
+
         this.userService.handleDeleteUser(id);
         return ResponseEntity.ok(null);
     }
